New Libraries InputManager and Jquery
The one and olny JQuery found its way in the project.
The InputManager helps to detect user inputs and makes it easy to work
with them.
I think the current state of the project is a good point for a Tag
because it has many funtionalitys to do many different games from here.